requirements.txt
tensorflow==2.16.1
pandas==1.3.4
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.7.2
pyyaml==6.0.1